+++ src/test/java/calculator/utils/DelimiterUtilsTest.java
@@ -1,7 +1,8 @@
-package calculator.validator;
+package calculator.utils;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import calculator.utils.DelimiterUtils;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -13,14 +14,14 @@ import org.junit.jupiter.params.provider.MethodSource;
  * @date : 2022/05/12 10:51 오후
  */
 @DisplayName("Validator:Delimiter")
-class DelimiterValidatorTest {
+class DelimiterUtilsTest {
 
     @ParameterizedTest
     @MethodSource
     @DisplayName("',' or ':' 구분자 포함 여부 판별")
     public void hasDelimiterTest(String given, boolean expected) {
         // When & Then
-        assertThat(DelimiterValidator.hasDelimiter(given)).isEqualTo(expected);
+        assertThat(DelimiterUtils.hasDelimiter(given)).isEqualTo(expected);
     }
 
     private static Stream hasDelimiterTest() {
@@ -37,7 +38,7 @@ class DelimiterValidatorTest {
     @DisplayName("“//”와 “\\n” 사이에 Custom 구분자 포함 여부 판별")
     public void hasCustomDelimiterTest(String given, boolean expected) {
         // When & Then
-        assertThat(DelimiterValidator.hasCustomDelimiter(given)).isEqualTo(expected);
+        assertThat(DelimiterUtils.hasCustomDelimiter(given)).isEqualTo(expected);
     }
 
     private static Stream hasCustomDelimiterTest() {
